Revert 199569 "Fix 64 bit compile errors flagged by conversions ..."
> Fix 64 bit compile errors flagged by conversions from size_t to uint32.
> 
> Using static_cast for now.
> 
> BUG=237249
> TBR=jam
> 
> Review URL: https://codereview.chromium.org/14914008

TBR=ananta@chromium.org

Review URL: https://codereview.chromium.org/14944007

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@199591 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: 5a3045130a5152f87f1107402ec20ca707fde7ea

+++ webkit/plugins/npapi/plugin_host.cc
@@ -482,7 +482,7 @@ static NPError PostURLNotify(NPP id,
       return NPERR_FILE_NOT_FOUND;
 
     buf = post_file_contents.c_str();
-    len = static_cast<uint32_t>(post_file_contents.size());
+    len = post_file_contents.size();
   }
 
   // The post data sent by a plugin contains both headers
@@ -988,10 +988,9 @@ NPError NPN_GetValueForURL(NPP id,
 
   // Allocate this using the NPAPI allocator. The plugin will call
   // NPN_Free to free this.
-  *value = static_cast<char*>(
-      NPN_MemAlloc(static_cast<uint32_t>(result.length()) + 1));
+  *value = static_cast<char*>(NPN_MemAlloc(result.length() + 1));
   base::strlcpy(*value, result.c_str(), result.length() + 1);
-  *len = static_cast<uint32_t>(result.length());
+  *len = result.length();
 
   return NPERR_NO_ERROR;
 }
